cc_sniffer/requirements.txt
scapy
PyYAML
prometheus-client
